+++ src/Requests/Application/Scan/Issue/CommentRequests.php
@@ -3,6 +3,8 @@
 namespace RIPS\Connector\Requests\Application\Scan\Issue;
 
 use GuzzleHttp\RequestOptions;
+use RIPS\Connector\Entities\Response;
+use RIPS\Connector\Exceptions\LibException;
 use RIPS\Connector\Requests\BaseRequest;
 
 class CommentRequests extends BaseRequest
@@ -30,14 +32,11 @@ class CommentRequests extends BaseRequest
      * @param int $scanId
      * @param int $issueId
      * @param array $queryParams
-     * @return \stdClass[]
+     * @return Response
      */
-    public function getAll($appId = null, $scanId = null, $issueId = null, array $queryParams = [])
+    public function getAll($appId, $scanId, $issueId, array $queryParams = [])
     {
-        $uri = is_null($appId)
-            ? "/applications/scans/issues/comments/all"
-            : $this->uri($appId, $scanId, $issueId);
-        $response = $this->client->get($uri, [
+        $response = $this->client->get($this->uri($appId, $scanId, $issueId), [
             'query' => $queryParams,
         ]);
 
@@ -52,7 +51,7 @@ class CommentRequests extends BaseRequest
      * @param int $issueId
      * @param int $commentId
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function getById($appId, $scanId, $issueId, $commentId, array $queryParams = [])
     {
@@ -71,7 +70,7 @@ class CommentRequests extends BaseRequest
      * @param int $issueId
      * @param array $input
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function create($appId, $scanId, $issueId, array $input, array $queryParams = [])
     {
@@ -90,7 +89,7 @@ class CommentRequests extends BaseRequest
      * @param int $scanId
      * @param int $issueId
      * @param array $queryParams
-     * @return void
+     * @return Response
      */
     public function deleteAll($appId, $scanId, $issueId, array $queryParams = [])
     {
@@ -98,7 +97,7 @@ class CommentRequests extends BaseRequest
             'query' => $queryParams,
         ]);
 
-        $this->handleResponse($response, null);
+        return $this->handleResponse($response);
     }
 
     /**
@@ -109,14 +108,18 @@ class CommentRequests extends BaseRequest
      * @param int $issueId
      * @param int $commentId
      * @param array $queryParams
-     * @return void
+     * @return Response
      */
     public function deleteById($appId, $scanId, $issueId, $commentId, array $queryParams = [])
     {
+        if (is_null($appId) || is_null($scanId) || is_null($issueId)) {
+            throw new LibException('appId, scanId, or issueId is null');
+        }
+
         $response = $this->client->delete($this->uri($appId, $scanId, $issueId, $commentId), [
             'query' => $queryParams,
         ]);
 
-        $this->handleResponse($response, null);
+        return $this->handleResponse($response);
     }
 }
