+++ src/Requests/Application/UploadRequests.php
@@ -2,6 +2,8 @@
 
 namespace RIPS\Connector\Requests\Application;
 
+use RIPS\Connector\Entities\Response;
+use RIPS\Connector\Exceptions\LibException;
 use RIPS\Connector\Requests\BaseRequest;
 
 class UploadRequests extends BaseRequest
@@ -29,7 +31,7 @@ class UploadRequests extends BaseRequest
      *
      * @param int|null $appId
      * @param array $queryParams
-     * @return \stdClass[]
+     * @return Response
      */
     public function getAll($appId = null, array $queryParams = [])
     {
@@ -46,7 +48,7 @@ class UploadRequests extends BaseRequest
      * @param int $appId
      * @param int $uploadId
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function getById($appId, $uploadId, array $queryParams = [])
     {
@@ -64,7 +66,7 @@ class UploadRequests extends BaseRequest
      * @param string $filename
      * @param string $filepath
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function create($appId, $filename, $filepath, array $queryParams = [])
     {
@@ -87,7 +89,7 @@ class UploadRequests extends BaseRequest
      *
      * @param int $appId
      * @param array $queryParams
-     * @return void
+     * @return Response
      */
     public function deleteAll($appId, array $queryParams = [])
     {
@@ -95,7 +97,7 @@ class UploadRequests extends BaseRequest
             'query' => $queryParams,
         ]);
 
-        $this->handleResponse($response, true);
+        return $this->handleResponse($response);
     }
 
     /**
@@ -104,14 +106,18 @@ class UploadRequests extends BaseRequest
      * @param int $appId
      * @param int $uploadId
      * @param array $queryParams
-     * @return void
+     * @return Response
      */
     public function deleteById($appId, $uploadId, array $queryParams = [])
     {
+        if (is_null($appId) || is_null($uploadId)) {
+            throw new LibException('appId or uploadId is null');
+        }
+
         $response = $this->client->delete($this->uri($appId, $uploadId), [
             'query' => $queryParams,
         ]);
 
-        $this->handleResponse($response, true);
+        return $this->handleResponse($response);
     }
 }
