+++ src/Requests/QuotaRequests.php
@@ -3,6 +3,8 @@
 namespace RIPS\Connector\Requests;
 
 use GuzzleHttp\RequestOptions;
+use RIPS\Connector\Entities\Response;
+use RIPS\Connector\Exceptions\LibException;
 use RIPS\Connector\Requests\Quota\AclRequests;
 
 class QuotaRequests extends BaseRequest
@@ -27,7 +29,7 @@ class QuotaRequests extends BaseRequest
      * Get all quotas
      *
      * @param array $queryParams
-     * @return \stdClass[]
+     * @return Response
      */
     public function getAll(array $queryParams = [])
     {
@@ -43,7 +45,7 @@ class QuotaRequests extends BaseRequest
      *
      * @param int $quotaId
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function getById($quotaId, array $queryParams = [])
     {
@@ -59,7 +61,7 @@ class QuotaRequests extends BaseRequest
      *
      * @param array $input
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function create(array $input, array $queryParams = [])
     {
@@ -77,7 +79,7 @@ class QuotaRequests extends BaseRequest
      * @param int $quotaId
      * @param array $input
      * @param array $queryParams
-     * @return \stdClass
+     * @return Response
      */
     public function update($quotaId, array $input, array $queryParams = [])
     {
@@ -93,7 +95,7 @@ class QuotaRequests extends BaseRequest
      * Delete all quotas
      *
      * @param array $queryParams
-     * @return void
+     * @return Response
      */
     public function deleteAll(array $queryParams = [])
     {
@@ -101,7 +103,7 @@ class QuotaRequests extends BaseRequest
             'query' => $queryParams,
         ]);
 
-        $this->handleResponse($response, true);
+        return $this->handleResponse($response);
     }
 
     /**
@@ -109,15 +111,19 @@ class QuotaRequests extends BaseRequest
      *
      * @param int $quotaId
      * @param array $queryParams
-     * @return void
+     * @return Response
      */
     public function deleteById($quotaId, array $queryParams = [])
     {
+        if (is_null($quotaId)) {
+            throw new LibException('quotaId is null');
+        }
+
         $response = $this->client->delete($this->uri($quotaId), [
             'query' => $queryParams,
         ]);
 
-        $this->handleResponse($response, true);
+        return $this->handleResponse($response);
     }
 
     /**
